packages/justhtml/justhtml-1.4.0.tar.gz/justhtml-1.4.0/tests/justhtml-tests/iframe_srcdoc.dat

#data
#iframe-srcdoc
<!DOCTYPE foo><p><table></table>
#errors
(1,14): unknown-doctype

#document
| <!DOCTYPE foo>
| <html>
|   <head>
|   <body>
|     <p>
|     <table>
